refactor: avoid api prefix for mock data
